+++ Versions/Temporary/Engine/Sources/AILogic/SoldierStates.cpp
@@ -973,7 +973,7 @@ void CSoldierMoveToState::Segment()
 			
 			if ( pUnit->GetNextCommand() == 0 )
 			{
-				const WORD wDir = pUnit->GetFrontDirection() == pUnit->GetDirection() ? wDirToPoint : wDirToPoint + 32768;
+				const WORD wDir = pUnit->GetFrontDirection() == pUnit->GetDirection() ? WORD(wDirToPoint) : WORD(int(wDirToPoint) + 32768);
 				theGroupLogic.UnitCommand( SAIUnitCmd( ACTION_COMMAND_GUARD, pUnit->GetCenterPlain(), wDir ), pUnit, false );
 				if ( bLongMove )
 					pUnit->SendAcknowledgement( NDb::ACK_MOVE_END, pUnit->GetPlayer() == theDipl.GetMyNumber() );
